try to have clocks as-fast-as-possible

+++ src/lib_utils/log.cpp
@@ -1,7 +1,6 @@
 #include "log.hpp"
 #include "clock.hpp"
 #include <cassert>
-#include <chrono>
 #include <ctime>
 #include <iostream>
 
@@ -42,7 +41,7 @@ std::string Log::getTime() {
 	const std::time_t t = std::time(nullptr);
 	std::tm tm = *std::localtime(&t);
 	auto const size = strftime(szOut, 255, "%Y/%m/%d %H:%M:%S", &tm);
-	return format("[%s][%s] ", std::string(szOut, size), g_DefaultClock->now() / IClock::Rate);
+	return format("[%s][%s] ", std::string(szOut, size), g_DefaultClock->now() / (double)IClock::Rate);
 }
 
 std::string Log::getColorBegin(Level level) {
